[IMP] Apply oca-addons-repo-template to project
The template used is this commit:
<https://github.com/coopiteasy/oca-addons-repo-template/commit/d28f504b7c2563a356f049d7a3e5b8700bf208d9>

At the time of writing, the 12.0 template for OCA is not yet finished,
see <https://github.com/OCA/oca-addons-repo-template/pull/117>. It was
used regardless, with some modifications.

The modifications are the additions of several tests in pre-commit, and
the bumping of flake8. These modifications are made because the source
material of this repo used these tools. The bumping of flake8 naively
fixes an unresolved problem from the linked PR.

Signed-off-by: Carmen Bianca Bakker <carmen@coopiteasy.be>